feat: gorm 2.0

+++ service/user_register_service.go
@@ -7,15 +7,15 @@ import (
 
 // UserRegisterService 管理用户注册服务
 type UserRegisterService struct {
-	Nickname        string `form:"nickname" json:"nickname" binding:"required,min=2,max=30"`
-	UserName        string `form:"user_name" json:"user_name" binding:"required,min=5,max=30"`
-	Password        string `form:"password" json:"password" binding:"required,min=8,max=40"`
+	Nickname string `form:"nickname" json:"nickname" binding:"required,min=2,max=30"`
+	UserName string `form:"user_name" json:"user_name" binding:"required,min=5,max=30"`
+	Password string `form:"password" json:"password" binding:"required,min=8,max=40"`
 }
 
 // valid 验证表单
 func (service *UserRegisterService) valid() *serializer.Response {
 
-	count := 0
+	count := int64(0)
 	model.DB.Model(&model.User{}).Where("nickname = ?", service.Nickname).Count(&count)
 	if count > 0 {
 		return &serializer.Response{
